LayoutTests/platform/mac/accessibility/scroll-to-visible-action-expected.txt










This tests that the scrollToMakeVisible accessibility action works as expected on the Mac.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Scroll to make visible is performed. The X position should be the same, but the Y position should have changed.
PASS newButtonX == buttonX is true
PASS newButtonY > buttonY is true
PASS successfullyParsed is true

TEST COMPLETE
